+++ lib/rubocop/cop/style/class_and_module_children.rb
@@ -23,12 +23,12 @@ module RuboCop
 
         NESTED_MSG = 'Use nested module/class definitions instead of ' \
                      'compact style.'.freeze
-
         COMPACT_MSG = 'Use compact module/class definition instead of ' \
                       'nested style.'.freeze
 
         def on_class(node)
-          _name, _superclass, body = *node
+          _name, superclass, body = *node
+          return if superclass
           check_style(node, body)
         end